+++ akshare/__init__.py
@@ -2621,9 +2621,10 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.12.6 fix: fix fund_open_fund_info_em interface
 1.12.7 fix: fix option_gfex_daily interface
 1.12.8 fix: fix fund_open_fund_info_em interface
+1.12.9 add: add fund_individual_basic_info_xq interface
 """
 
-__version__ = "1.12.8"
+__version__ = "1.12.9"
 __author__ = "AKFamily"
 
 import sys
@@ -5038,4 +5039,15 @@ from akshare.futures.futures_daily_bar import (
     get_futures_daily,
     get_ine_daily,
     get_gfex_daily,
+)
+
+"""
+雪球基金数据
+"""
+from akshare.fund.fund_xq import (
+    fund_individual_basic_info_xq,
+    fund_individual_achievement_xq,
+    fund_individual_analysis_xq,
+    fund_individual_profit_probability_xq,
+    fund_individual_detail_info_xq,
 )