+++ docs/changelog.md
@@ -70,6 +70,16 @@
 
 ## 更新说明详情
 
+1.12.9 add: add fund_individual_basic_info_xq interface
+
+    1. 新增 fund_individual_basic_info_xq 接口
+    2. 新增 fund_individual_achievement_xq 接口
+    3. 新增 fund_individual_analysis_xq 接口
+    4. 新增 fund_individual_profit_probability_xq 接口
+    5. 新增 fund_individual_detail_info_xq 接口
+    6. 新增 get_receipt 接口
+    7. 新增 reits_realtime_em 接口
+
 1.12.8 fix: fix fund_open_fund_info_em interface
 
     1. 修复 fund_open_fund_info_em 接口
@@ -3293,6 +3303,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.12.9 add: add fund_individual_basic_info_xq interface
 
 1.12.8 fix: fix fund_open_fund_info_em interface